+++ playground/main.ts
@@ -10,7 +10,10 @@ import {
   RenderResult,
   RenderResultSuccessNonOIDCSession,
 } from '../src/types';
-import { OktaSignInAPI as OktaSignInAPIV3 } from '../src/v3/src/types';
+import {
+  OktaSignInAPI as OktaSignInAPIV3,
+  WidgetOptions as WidgetOptionsV3,
+} from '../src/v3/src/types';
 import { assertNoEnglishLeaks } from '../playground/LocaleUtils';
 import Util from '../src/util/Util';
 import { addAfterTransformHooks, addHookOptions } from './hooks';
@@ -63,7 +66,7 @@ const renderPlaygroundWidget = (options: WidgetOptions & { assertNoEnglishLeaks?
   // Okta-hosted widget page has this value set for CSP
   window.cspNonce = 'playground';
 
-  addHookOptions(options);
+  addHookOptions(options as any as WidgetOptionsV3);
 
   createWidgetInstance(options);
