--- src/v3/src/types/registrationSchema.ts
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2022-present, Okta, Inc. and/or its affiliates. All rights reserved.
- * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
- *
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *
- * See the License for the specific language governing permissions and limitations under the License.
- */
-
-import { Input } from '@okta/okta-auth-js';
-import { IdxOption } from '@okta/okta-auth-js/types/lib/idx/types/idx-js';
-
-// Utility to modify interfaces / types
-type Modify<T, R> = Omit<T, keyof R> & R;
-
-export type RegistrationElementSchema = Modify<Input, {
-  'label-top'?: boolean;
-  placeholder?: string;
-  'data-se'?: string;
-  options?: IdxOption[] | Record<string, string>;
-  sublabel?: string;
-  wide?: boolean;
-}>;
